refactor: remove External Tools feature
Delete prof tools benchstat/qcachegrind, hub External Tools menu, engine/tools wrappers, and related app/intent/CLI wiring.
