+++ cli/constants_vars.go
@@ -1,6 +1,6 @@
 package cli
 
-// Shared flag names for track and tools commands.
+// Shared flag names for track and collect commands.
 const (
 	tuiPageSize          = 20
 	minTagsForComparison = 2
@@ -10,14 +10,3 @@ const (
 	benchNameFlag  = "bench-name"
 	tagFlag        = "tag"
 )
-
-// toolsFlags holds prof tools subcommand flags.
-type toolsFlags struct {
-	baseline      string
-	current       string
-	benchmarkName string
-	tag           string
-	profileType   string
-}
-
-var toolsGlobal toolsFlags
