--- engine/tools/benchstats/doc.go
@@ -1,2 +1,0 @@
-// Package benchstats wraps the benchstat CLI to compare two collected benchmark text files.
-package benchstats
