--- engine/tools/qcachegrind/doc.go
@@ -1,2 +1,0 @@
-// Package qcachegrind generates callgrind output from a collected binary profile and launches qcachegrind.
-package qcachegrind
