+++ internal/app/app_test.go
@@ -20,11 +20,6 @@ type stubTracker struct{}
 func (stubTracker) RunTrackAuto(_ TrackOptions) error   { return nil }
 func (stubTracker) RunTrackManual(_ TrackOptions) error { return nil }
 
-type stubTools struct{}
-
-func (stubTools) RunBenchStats(_, _, _ string) error  { return nil }
-func (stubTools) RunQcacheGrind(_, _, _ string) error { return nil }
-
 func TestWithDefaultsFillsAgentWhenNil(t *testing.T) {
 	s := &Services{}
 	out := s.WithDefaults()
@@ -44,7 +39,7 @@ func (stubSetup) CreateTemplate() error { return nil }
 func TestWithDefaultsNilReceiver(t *testing.T) {
 	var s *Services
 	out := s.WithDefaults()
-	if out == nil || out.Runner == nil || out.Collect == nil || out.Tracker == nil || out.Tools == nil || out.Setup == nil || out.Config == nil {
+	if out == nil || out.Runner == nil || out.Collect == nil || out.Tracker == nil || out.Setup == nil || out.Config == nil {
 		t.Fatalf("expected all fields set: %#v", out)
 	}
 }
@@ -52,7 +47,7 @@ func TestWithDefaultsNilReceiver(t *testing.T) {
 func TestWithDefaultsAllNilFieldsNonNilReceiver(t *testing.T) {
 	s := &Services{}
 	out := s.WithDefaults()
-	if out.Collect == nil || out.Tracker == nil || out.Tools == nil || out.Setup == nil {
+	if out.Collect == nil || out.Tracker == nil || out.Setup == nil {
 		t.Fatalf("expected defaults for every slot: %#v", out)
 	}
 }
@@ -73,11 +68,10 @@ func TestWithDefaultsPreservesAllNonNil(t *testing.T) {
 	s := &Services{
 		Collect: stubCollect{},
 		Tracker: stubTracker{},
-		Tools:   stubTools{},
 		Setup:   stubSetup{},
 	}
 	out := s.WithDefaults()
-	if out.Collect != s.Collect || out.Tracker != s.Tracker || out.Tools != s.Tools || out.Setup != s.Setup {
+	if out.Collect != s.Collect || out.Tracker != s.Tracker || out.Setup != s.Setup {
 		t.Fatal("WithDefaults replaced a non-nil dependency")
 	}
 }
@@ -125,17 +119,6 @@ func TestDefaultTrackerInvalidFormat(t *testing.T) {
 	}
 	if d.Tracker.RunTrackManual(opts) == nil {
 		t.Fatal("expected invalid format error")
-	}
-}
-
-func TestDefaultToolsEarlyErrors(t *testing.T) {
-	t.Chdir(t.TempDir())
-	d := Default()
-	if d.Tools.RunBenchStats("a", "b", "c") == nil {
-		t.Fatal("expected error when bench dir missing")
-	}
-	if d.Tools.RunQcacheGrind("tag", "bench", "cpu") == nil {
-		t.Fatal("expected error when profile binary missing")
 	}
 }
