+++ internal/intent/doc.go
@@ -13,8 +13,6 @@
 //   - KindSetup / SetupIntent → Config.CreateDefaultFile (deprecated alias)
 //   - KindConfigCreate / ConfigCreateIntent → Config.CreateDefaultFile
 //   - KindConfigSave / ConfigSaveIntent → Config.Save
-//   - KindToolsBenchstat / ToolsBenchstatIntent → Tools.RunBenchStats
-//   - KindToolsQcachegrind / ToolsQcachegrindIntent → Tools.RunQcacheGrind
 //
 // New workflows: add a Kind constant, an entry to AllKinds, a new file with types implementing Executable,
 // and wire cli or cli/tui to construct the intent after Survey prompts.
