+++ internal/tui/hub.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/AlexsanderHamir/prof/internal/workspace"
-
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
 )
@@ -22,8 +20,6 @@ const (
 	MainCollect
 	// MainCompare runs interactive compare between two tags.
 	MainCompare
-	// MainTools opens the tools submenu (benchstat, qcachegrind).
-	MainTools
 	// MainSetup is a deprecated alias for MainConfig.
 	MainSetup
 	// MainConfig creates prof.json when missing.
@@ -60,7 +56,6 @@ func newHubModel() *hubModel {
 		items: []mainItem{
 			{"Collect Profiles", MainCollect},
 			{"Compare Runs", MainCompare},
-			{"External Tools", MainTools},
 			{"Create Configuration File", MainConfig},
 			{"Documentation Site", MainDocs},
 			{"Quit", MainQuit},
@@ -151,7 +146,6 @@ func (m *hubModel) View() string {
 		b.WriteString(helpStyle.Render(
 			"Collect Profiles: run benchmarks and store output under bench/<tag>/.\n" +
 				"Compare Runs: diff two saved runs function-by-function; optional regression limits from prof.json or CLI.\n" +
-				fmt.Sprintf("External Tools: run %s or %s.\n", workspace.ToolNameBenchstat, workspace.ToolNameQcachegrind) +
 				"Create Configuration File: writes prof.json and prof.json.example beside go.mod.\n" +
 				"Press ? again to hide this help.",
 		))
